requirements.txt
discord.py
python-dotenv
aiosqlite
openai
watchfiles
PyNaCl
